Se valida las publicaciones

+++ RedSocialPersistencia/src/main/java/daos/PublicacionDAO.java
@@ -15,6 +15,7 @@ import com.mongodb.client.result.UpdateResult;
 import dominio.Anclada;
 import dominio.Comun;
 import dominio.Publicacion;
+import excepciones.MongoDBException;
 import interfaces.IPublicacionDAO;
 import java.util.ArrayList;
 import java.util.List;
@@ -39,11 +40,16 @@ public class PublicacionDAO implements IPublicacionDAO {
     }
 
     @Override
-    public Comun registrarPublicacionComun(Comun publicacionComun) {
+    public Comun registrarPublicacionComun(Comun publicacionComun) throws MongoDBException {
         try {
+            List<String> errores = validarPublicacionComun(publicacionComun);
+            if (!errores.isEmpty()) {
+                String mensajeError = String.join(", ", errores);
+                throw new MongoDBException("La publicación común no es válida: " + mensajeError);
+            }
             COLECCION.insertOne(publicacionComun);
             return publicacionComun;
-        } catch (Exception e) {
+        } catch (MongoDBException e) {
             System.out.println("Error al registrar la publicación comun: " + e.getMessage());
             return null;
         }
@@ -52,9 +58,15 @@ public class PublicacionDAO implements IPublicacionDAO {
     @Override
     public Anclada registrarPublicacionAnclada(Anclada anclada) {
         try {
+            List<String> errores = validarPublicacionAnclada(anclada);
+
+            if (!errores.isEmpty()) {
+                String mensajeError = String.join(", ", errores);
+                throw new MongoDBException("La publicación anclada no es válida: " + mensajeError);
+            }
             COLECCION.insertOne(anclada);
             return anclada;
-        } catch (Exception e) {
+        } catch (MongoDBException e) {
             System.out.println("Error al registrar la publicación }{ñljk.anclada0-: " + e.getMessage());
             return null;
         }
@@ -63,6 +75,11 @@ public class PublicacionDAO implements IPublicacionDAO {
     @Override
     public Comun editarPublicacionComun(Comun publicacionComun) {
         try {
+            List<String> errores = validarPublicacionComun(publicacionComun);
+            if (!errores.isEmpty()) {
+                String mensajeError = String.join(", ", errores);
+                throw new MongoDBException("La publicación común no es válida: " + mensajeError);
+            }
             ObjectId publicacionId = publicacionComun.getId();
             Bson filtro = Filters.eq("_id", publicacionId);
             Bson update = new Document("$set", publicacionComun);
@@ -75,7 +92,7 @@ public class PublicacionDAO implements IPublicacionDAO {
                 System.out.println("No se encontró la publicación comun para editar");
                 return null;
             }
-        } catch (Exception e) {
+        } catch (MongoDBException e) {
             System.out.println("Error al editar la publicación comun: " + e.getMessage());
             return null;
         }
@@ -92,7 +109,7 @@ public class PublicacionDAO implements IPublicacionDAO {
             }
             cursor.close();
             return publicaciones;
-        } catch (Exception e) {
+        } catch (MongoDBException e) {
             System.out.println("Error al consultar las publicaciones: " + e.getMessage());
             return null;
         }
@@ -101,6 +118,11 @@ public class PublicacionDAO implements IPublicacionDAO {
     @Override
     public boolean eliminarPublicacion(Publicacion publicacion) {
         try {
+            List<String> errores = validarPublicacion(publicacion);
+            if (!errores.isEmpty()) {
+                String mensajeError = String.join(", ", errores);
+                throw new MongoDBException("La publicación no es válida: " + mensajeError);
+            }
             ObjectId publicacionId = publicacion.getId();
             Bson filtro = Filters.eq("_id", publicacionId);
             DeleteResult resultado = COLECCION.deleteOne(filtro);
@@ -112,7 +134,7 @@ public class PublicacionDAO implements IPublicacionDAO {
                 System.out.println("No se encontró la publicación para eliminar");
                 return false;
             }
-        } catch (Exception e) {
+        } catch (MongoDBException e) {
             System.out.println("Error al eliminar la publicación: " + e.getMessage());
             return false;
         }
@@ -121,14 +143,86 @@ public class PublicacionDAO implements IPublicacionDAO {
     @Override
     public boolean existePublicacion(Publicacion publicacion) {
         try {
+            List<String> errores = validarPublicacion(publicacion);
+            if (!errores.isEmpty()) {
+                String mensajeError = String.join(", ", errores);
+                throw new MongoDBException("La publicación no es válida: " + mensajeError);
+            }
             ObjectId publicacionId = publicacion.getId();
             Bson filtro = Filters.eq("_id", publicacionId);
             long count = COLECCION.countDocuments(filtro);
             return count > 0;
-        } catch (Exception e) {
+        } catch (MongoDBException e) {
             System.out.println("Error al verificar la existencia de la publicación: " + e.getMessage());
             return false;
         }
     }
 
+    //Validaciones 
+    private List<String> validarPublicacion(Publicacion publicacion) {
+        List<String> errores = new ArrayList<>();
+
+        // Validar los atributos de la publicación común
+        if (publicacion.getFechaHoraCreacion() == null) {
+            errores.add("El campo FechaHoraCreacion es requerido");
+        }
+
+        if (publicacion.getTitulo() == null || publicacion.getTitulo().isEmpty()) {
+            errores.add("El campo Titulo es requerido");
+        }
+
+        if (publicacion.getContenido() == null || publicacion.getContenido().isEmpty()) {
+            errores.add("El campo Contenido es requerido");
+        }
+
+        return errores;
+    }
+
+    private List<String> validarPublicacionComun(Comun publicacionComun) {
+        List<String> errores = new ArrayList<>();
+
+        // Validar los atributos de la publicación común
+        if (publicacionComun.getFechaHoraCreacion() == null) {
+            errores.add("El campo FechaHoraCreacion es requerido");
+        }
+
+        if (publicacionComun.getTitulo() == null || publicacionComun.getTitulo().isEmpty()) {
+            errores.add("El campo Titulo es requerido");
+        }
+
+        if (publicacionComun.getContenido() == null || publicacionComun.getContenido().isEmpty()) {
+            errores.add("El campo Contenido es requerido");
+        }
+
+        // Validar atributos específicos de la clase Comun
+        if (publicacionComun.getUsuario() == null) {
+            errores.add("El campo Usuario es requerido");
+        }
+
+        return errores;
+    }
+
+    private List<String> validarPublicacionAnclada(Anclada anclada) {
+        List<String> errores = new ArrayList<>();
+
+        // Validar los atributos de la publicación anclada
+        if (anclada.getFechaHoraCreacion() == null) {
+            errores.add("El campo FechaHoraCreacion es requerido");
+        }
+
+        if (anclada.getTitulo() == null || anclada.getTitulo().isEmpty()) {
+            errores.add("El campo Titulo es requerido");
+        }
+
+        if (anclada.getContenido() == null || anclada.getContenido().isEmpty()) {
+            errores.add("El campo Contenido es requerido");
+        }
+
+        // Validar atributos específicos de la clase Anclada
+        if (anclada.getAdmin() == null) {
+            errores.add("El campo Admin es requerido");
+        }
+
+        return errores;
+    }
 }
